+++ app/src/main/java/com/bodan/maplecalendar/presentation/broadcastreceiver/MyAlarmReceiver.kt
@@ -16,7 +16,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
 import com.bodan.maplecalendar.R
-import com.bodan.maplecalendar.data.EventListReader
+import com.bodan.maplecalendar.data.util.EventListReader
 import com.bodan.maplecalendar.presentation.views.SplashActivity
 import com.google.android.material.internal.ManufacturerUtils
 import kotlinx.coroutines.CoroutineScope
